requirements.txt
cffi==1.14.3
cryptography==3.1.1
Django==2.2
django-crispy-forms==1.9.2
django-environ==0.4.5
django-markdownx==3.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-taggit==1.3.0
django-uuslug==1.2.0
importlib-metadata==2.0.0
Markdown==3.2.2
Pillow==7.2.0
pycparser==2.20
PyMySQL==0.10.1
python-slugify==4.0.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
zipp==3.2.0
